Set NetShield value in local-agent on connect [VPNAND-831].
UserData.getNetShieldProtocol returns value that depends on the logged
in user value. VpnBackend only observes for changes to the value made
via UserData.setNetShieldProtocol and its possible for it to have a
stale value (e.g. from before the user logged in).
This is a quick fix to set the value after connecting.