Fix WakeLock autorelease [VPNAND-1018].
Wake locks held for connecting and fallback logic are acquired with
timeout to prevent using too much power if the job takes too much time.
If the lock is released by timeout it should not be released manually
any more, otherwise this will count as an additional release. Older
Android versions throw an exception when this happens but even on newer
versions this skews the reference counting.

With this change the wake locks are only released if still held. There's
a separate wake lock for connect and fallback so that there's no need to
rely on the reference counting mechanism.